style: improve folder and file structures
