+++ client/app/components/project-screen/sidebar-panel/activity-panel/activity-bar.tsx
