+++ client/app/components/project-screen/sidebar-panel/chat-panel/new-message-input.tsx
