+++ client/app/components/project-screen/sidebar-panel/chat-panel/no-chat-items-image.tsx
