+++ client/app/components/ui/header.tsx
@@ -1,6 +1,6 @@
-import CreateProjectBtn from "@/components/home/create-project";
+import CreateProjectBtn from "@/components/home-screen/create-project";
+import JoinProjectBtn from "@/components/home-screen/join-project";
 import Logo from "./logo";
-import JoinProjectBtn from "@/components/home/join-project";
 
 export default function Header() {
   return (
