+++ client/app/page.tsx
@@ -1,7 +1,7 @@
-import CheckWelcomeScreen from "@/components/home/check-welcome-screen";
-import HomeScreenLoading from "./components/home/home-screen-loading";
-import WelcomeScreen from "./components/home/welcome-screen";
-import HomeScreen from "./components/home/home-screen";
+import CheckWelcomeScreen from "@/components/welcome-screen/check-welcome-screen";
+import HomeScreenLoading from "./components/home-screen/home-screen-loading";
+import WelcomeScreen from "./components/welcome-screen/welcome-screen";
+import HomeScreen from "./components/home-screen/home-screen";
 
 export default function Home() {
   return (
