+++ client/app/project/[projectId]/page.tsx
@@ -1,7 +1,7 @@
-import CheckProjectAvailability from "@/components/project/check-project-availability";
-import ProjectPageLoading from "@/app/components/project/project-page-loading";
-import ProjectPageIsNotAvailable from "@/app/components/project/project-page-not-available";
-import ProjectPage from "@/app/components/project/project-page";
+import ProjectPage from "@/app/components/project-screen/project-page";
+import ProjectPageLoading from "@/app/components/project-screen/project-page-loading";
+import ProjectPageIsNotAvailable from "@/app/components/project-screen/project-page-not-available";
+import CheckProjectAvailability from "@/components/project-screen/check-project-availability";
 
 export default function page() {
   return (
