+++ client/components/home-screen/card/card-connected-user-count.tsx
