+++ client/components/home-screen/card/card-open-project-btn.tsx
