+++ client/components/home-screen/card/card-owner.tsx
