+++ client/components/home-screen/join-project.tsx
