+++ client/components/project-screen/activity-bar/activity-bar-btn.tsx
