+++ client/components/project-screen/check-project-availability.tsx
