+++ client/components/project-screen/code-editor/code-area.tsx
