+++ client/components/project-screen/code-editor/editor-status-bar.tsx
