+++ client/components/project-screen/goback-on-home-page-btn.tsx
@@ -3,7 +3,7 @@
 import Link from "next/link";
 import { Button } from "../ui/button";
 
-export default function GobackOnHommePage () {
+export default function GobackOnHomePageBtn() {
   return (
     <div>
       <Link href="/">
