+++ client/components/project-screen/mount-project-socket-events.tsx
