+++ client/components/project-screen/sidebar-panel/file-explorer/rename-project-item.tsx
