+++ client/components/welcome-screen/check-welcome-screen.tsx
