+++ client/components/welcome-screen/get-started-now-btn.tsx
